# mpt: support put in batches