When updating MPT via multiple `Put` there is substantial overhead as intermediate nodes are updated on each put. However we care only about the state on block boundaries.
This PR adds support for put in batches which doesn't calculate hashes for intermediate nodes.
There is no meaningful difference on benchmark of 4 nodes, but single node seems to benefit from this.
Restore times are also much less (40% gain for 10k tx per block, and 15% for 1 tx per block), same for memory consumption:
```
# master, 50x10000
TotalAlloc (MiB): 16116
Mallocs: 224004004
Frees: 206607635
PauseTotalNs: 2669760
GCCPUFraction:: 0.03410
./neogo db restore --unittest --in ./chain50x10000.acc --dump before  62,49s user 0,88s system 170% cpu 37,103 total

# PR, 50x10000
TotalAlloc (MiB): 9654
Mallocs: 151838952
Frees: 139728189
PauseTotalNs: 16819244
GCCPUFraction:: 0.02739
./neogo db restore --unittest --in ./chain50x10000.acc --dump after  38,05s user 0,74s system 161% cpu 23,979 total

# master, 10000x1
TotalAlloc (MiB): 17977
Mallocs: 247770125
Frees: 226787985
PauseTotalNs: 3636223
GCCPUFraction:: 0.02953
./neogo db restore --unittest --in ./chain10000x1.acc --dump before  73,64s user 1,36s system 177% cpu 42,225 total

# PR, 10000x1
TotalAlloc (MiB): 13838
Mallocs: 194745452
Frees: 172326918
PauseTotalNs: 2780536
GCCPUFraction:: 0.03022
./neogo db restore --unittest --in ./chain10000x1.acc --dump after  59,01s user 1,46s system 177% cpu 34,030 total
```

States compared via `scripts/compare_dumps.go` are the same.